+++ code/sys.ui/react.css/src/style.Color/t.ts
@@ -1,0 +1,3 @@
+import type { t } from '../common';
+
+export {};
